+++ light-type.h
@@ -12,11 +12,11 @@ class LightType
     virtual void loop() {};
     virtual void setColor1(RGB color) {};
     virtual void setColor2(RGB color) {};
-    virtual void setParam(int param) {};
+    virtual void setParam(uint16_t param) {};
 
     virtual RGB getColor1() {return {0,0,0};};
     virtual RGB getColor2() {return {0,0,0};};
-    virtual int getParam() {return 0;};
+    virtual uint16_t getParam() {return 0;};
 
     Adafruit_NeoPixel* _strip;
 };
